requirements.txt
langchain>=0.1.20
sentence-transformers>=2.7
scikit-learn>=1.5
feedparser>=6.0.11
newspaper3k>=0.2.8
python-dotenv>=1.0.1
PyYAML>=6.0.1
ollama-python==0.1.2
